+++ src/app/workspace/search/expert-search/expert-search.component.spec.ts
@@ -1,6 +1,7 @@
 import { Component, DebugElement, OnInit, ViewChild } from '@angular/core';
-import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatDialogModule } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { By } from '@angular/platform-browser';
@@ -53,11 +54,12 @@ describe('ExpertSearchComponent', () => {
                 TestHostComponent
             ],
             imports: [
-                FormsModule,
-                ReactiveFormsModule,
                 BrowserAnimationsModule,
+                FormsModule,
+                MatDialogModule,
                 MatFormFieldModule,
-                MatInputModule
+                MatInputModule,
+                ReactiveFormsModule
             ],
             providers: [
                 {